requirements.txt
aiofiles==24.1.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.0
asyncpg==0.29.0
bcrypt==4.0.1
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
environs==11.0.0
et-xmlfile==1.1.0
fastapi==0.115.0
fastapi-cli==0.0.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
numpy==2.1.1
openpyxl==3.1.5
orjson==3.10.7
packaging==24.1
pandas==2.2.3
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
rich==13.9.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.38.6
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
uvicorn==0.31.0
watchfiles==0.24.0
websockets==13.1
